Update the Math.Random use in the computerPlay method
Before, the Math.Random returned a value between 1 and 100.
This value was then used to define the choice of the computer using
percentage. But this solution was neither clean nor correct.

Now, the Math.Random returns a value between 0 and 2, that is used to
return the value at this index in an array containing our 3 possibilities.